+++ src/main/java/net/geforcemods/securitycraft/blockentities/ReinforcedLecternBlockEntity.java
@@ -1,0 +1,139 @@
+package net.geforcemods.securitycraft.blockentities;
+
+import java.util.EnumMap;
+import java.util.Map;
+
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.api.ICustomizable;
+import net.geforcemods.securitycraft.api.IModuleInventory;
+import net.geforcemods.securitycraft.api.IOwnable;
+import net.geforcemods.securitycraft.api.Option;
+import net.geforcemods.securitycraft.api.Option.BooleanOption;
+import net.geforcemods.securitycraft.api.Owner;
+import net.geforcemods.securitycraft.inventory.ReinforcedLecternMenu;
+import net.geforcemods.securitycraft.misc.ModuleType;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.NonNullList;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.Connection;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.entity.LecternBlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+
+public class ReinforcedLecternBlockEntity extends LecternBlockEntity implements IOwnable, IModuleInventory, ICustomizable {
+	private Owner owner = new Owner();
+	private NonNullList<ItemStack> modules = NonNullList.withSize(getMaxNumberOfModules(), ItemStack.EMPTY);
+	private Map<ModuleType, Boolean> moduleStates = new EnumMap<>(ModuleType.class);
+	private BooleanOption lockPage = new BooleanOption("lockPage", false);
+
+	public ReinforcedLecternBlockEntity(BlockPos pos, BlockState state) {
+		super(pos, state);
+	}
+
+	@Override
+	public void load(CompoundTag tag) {
+		super.load(tag);
+		owner.load(tag);
+		modules = readModuleInventory(tag);
+		moduleStates = readModuleStates(tag);
+		readOptions(tag);
+	}
+
+	@Override
+	public void saveAdditional(CompoundTag tag) {
+		super.saveAdditional(tag);
+
+		if (owner != null)
+			owner.save(tag, needsValidation());
+
+		writeModuleInventory(tag);
+		writeModuleStates(tag);
+		writeOptions(tag);
+	}
+
+	@Override
+	public CompoundTag getUpdateTag() {
+		return saveWithoutMetadata();
+	}
+
+	@Override
+	public ClientboundBlockEntityDataPacket getUpdatePacket() {
+		return ClientboundBlockEntityDataPacket.create(this);
+	}
+
+	@Override
+	public void onDataPacket(Connection net, ClientboundBlockEntityDataPacket packet) {
+		super.onDataPacket(net, packet);
+		handleUpdateTag(packet.getTag());
+	}
+
+	@Override
+	public void handleUpdateTag(CompoundTag tag) {
+		load(tag);
+	}
+
+	@Override
+	public Owner getOwner() {
+		return owner;
+	}
+
+	@Override
+	public void setOwner(String uuid, String name) {
+		owner.set(uuid, name);
+		setChanged();
+	}
+
+	@Override
+	public ModuleType[] acceptedModules() {
+		return new ModuleType[] {
+				ModuleType.ALLOWLIST
+		};
+	}
+
+	@Override
+	public Option<?>[] customOptions() {
+		return new Option[] {
+				lockPage
+		};
+	}
+
+	public boolean isPageLocked() {
+		return lockPage.get();
+	}
+
+	@Override
+	public NonNullList<ItemStack> getInventory() {
+		return modules;
+	}
+
+	@Override
+	public boolean isModuleEnabled(ModuleType module) {
+		return hasModule(module) && moduleStates.get(module) == Boolean.TRUE; //prevent NPE
+	}
+
+	@Override
+	public void toggleModuleState(ModuleType module, boolean shouldBeEnabled) {
+		moduleStates.put(module, shouldBeEnabled);
+	}
+
+	@Override
+	public BlockEntityType<?> getType() {
+		return SCContent.REINFORCED_LECTERN_BLOCK_ENTITY.get();
+	}
+
+	@Override
+	public AbstractContainerMenu createMenu(int containerId, Inventory playerInventory, Player player) {
+		return new ReinforcedLecternMenu(containerId, this);
+	}
+
+	@Override
+	public Component getDisplayName() {
+		return Component.translatable(SCContent.REINFORCED_LECTERN.get().getDescriptionId());
+	}
+}