+++ src/main/java/net/geforcemods/securitycraft/items/UniversalBlockReinforcerItem.java
@@ -26,6 +26,7 @@ import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.LecternBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.neoforged.neoforge.network.NetworkHooks;
 
@@ -83,6 +84,8 @@ public class UniversalBlockReinforcerItem extends Item {
 
 					if (be instanceof Container container)
 						container.clearContent();
+					else if (be instanceof LecternBlockEntity lectern)
+						lectern.clearContent();
 				}
 
 				level.setBlockAndUpdate(pos, convertedState);
@@ -97,11 +100,11 @@ public class UniversalBlockReinforcerItem extends Item {
 				}
 
 				stack.hurtAndBreak(1, player, p -> p.broadcastBreakEvent(p.getUsedItemHand()));
-				return false;
+				return true;
 			}
 		}
 
-		return true;
+		return false;
 	}
 
 	public static boolean isReinforcing(ItemStack stack) {